prototype complete.
input is kind of clunky, but it works - you have an actions file
containing recipes you want oreganizer to know about, a goals file that
contains a set of things that you want to have, and a resources file
that contains things you already have.

there currently isn't any memory of any kind. the only way to interact
with the program is through the input files. if you build something and
want to take it off the list of things you built, you have to manually
add it to the resource file.