change over to using sparse matrices.
new scheme will require a matrix with a number of columns linear in the
number of materials and edges and a number of rows linear in the number
of edges in the dependency graph. This is unfortunately dogawful,
because, really, O(ve + e^2)? Eurgh. Thankfully, sparse matrices to the
rescue! An overwhelmingly vast majority of that matrix is zero because
the our linear equations are never going to have more than eight or ten
nonzero coefficients each while we have thousands of equations. Which
means our matrix is lolsparse and our complexities are back into the
realm of being linear in the number of edges. Yay math!