finally got the code clean enough to move on
added comments /everywhere/. ran into an issue while I was explaining
that with requires dependencies, too; there's no way to encode
conditionals in a programming problem like this (because conditionals
aren't convex), so we're going to have to do something silly. Namely, a
really kludgey relative of column generation where we run the solver,
check to see if we need to make anything with a requirement, and if we
do add an endogenous "want" and go. I feel like I'm ready to start
making that change I discussed a couple commits ago where I need a step
in between needs and wants.